requirements.txt
# Django Core
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Authentication
djangorestframework-simplejwt>=5.2.0

# Database
mysqlclient>=2.1.0

# OCR Libraries
pytesseract>=0.3.10
Pillow>=10.0.0
PyPDF2>=3.0.0

# Google Vision API (Optional)
google-cloud-vision>=3.4.0

# AI Integration
openai>=1.0.0

# Environment Variables
python-dotenv>=1.0.0

# Development Tools
django-extensions>=3.2.0

# Security
django-environ>=0.11.0
